+++ src/lib.rs
@@ -2,34 +2,6 @@ use core::cmp::Ordering;
 use core::str::FromStr;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
-pub trait Broadcast {
-	type Output;
-
-	fn broadcast(&self) -> Self::Output;
-}
-
-impl Broadcast for Netv4Addr {
-	type Output = Ipv4Addr;
-
-	fn broadcast(&self) -> Ipv4Addr {
-		let netmask: u32 = self.mask().clone().into();
-		let network: u32 = self.addr().clone().into();
-		let broadcast: u32 = network | !netmask;
-		broadcast.into()
-	}
-}
-
-impl Broadcast for NetAddr {
-	type Output = Option<IpAddr>;
-
-	fn broadcast(&self) -> Self::Output {
-		match self {
-			NetAddr::V4(netaddr) => Some(IpAddr::from(netaddr.broadcast())),
-			_ => None,
-		}
-	}
-}
-
 trait Merge {
 	type Output;
 
@@ -345,6 +317,9 @@ impl PartialOrd for Netv6Addr {
 		}
 	}
 }
+
+mod broadcast;
+pub use broadcast::*;
 
 mod mask;
 pub use mask::*;
